requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-environ>=0.11
django-cors-headers>=4.2
drf-spectacular>=0.26
PyYAML>=6.0
